Python Scripts
A series of Python scripts using 2.x and 3.x.  

Python 3.5:
Dataentry.py creates a UI for the input of basketball stats.  

Sale.py is a very basic UI that has a stock body for a webpage, but new ones can be created and called.  When run, the webpage is created.

Python 2:

CopyMod.py is a UI that allows a user to choose a source location and a destination location and copy new or modified into the new location.  Uses SQLITE3 and wxPython.

folderbrowser.py does just that.